+++ semcore/wizard/src/Wizard.jsx
@@ -8,13 +8,17 @@ import i18nEnhance from '@semcore/utils/lib/enhances/i18nEnhance';
 import { localizedMessages } from './translations/__intergalactic-dynamic-locales';
 import { ScreenReaderOnly } from '@semcore/utils/lib/ScreenReaderOnly';
 import { setFocus } from '@semcore/utils/lib/use/useFocusLock';
+import uniqueIDEnhancement from '@semcore/utils/lib/uniqueID';
+import Button from '@semcore/button';
+import ArrowRight from '@semcore/icon/ArrowRight/m';
+import ArrowLeft from '@semcore/icon/ArrowLeft/m';
 
 import style from './style/wizard.shadow.css';
 
 class WizardRoot extends Component {
   static displayName = 'Wizard';
   static style = style;
-  static enhance = [i18nEnhance(localizedMessages)];
+  static enhance = [i18nEnhance(localizedMessages), uniqueIDEnhancement()];
   static defaultProps = {
     step: null,
     i18n: localizedMessages,
@@ -23,13 +27,59 @@ class WizardRoot extends Component {
 
   _steps = new Map();
   modalRef = React.createRef();
+  contentRef = React.createRef();
+  state = { highlighted: null };
 
   getStepProps(props) {
     return {
       steps: this._steps,
       active: props.step === this.asProps.step,
+      id: `${this.asProps.uid}-step-${props.step}`,
     };
   }
+
+  getSidebarProps() {
+    return {
+      uid: this.asProps.uid,
+    };
+  }
+  getContentProps() {
+    return {
+      uid: this.asProps.uid,
+      step: this.asProps.step,
+      ref: this.contentRef,
+    };
+  }
+  getStepBackProps() {
+    return {
+      getI18nText: this.asProps.getI18nText,
+      step: this.asProps.step,
+    };
+  }
+  getStepNextProps() {
+    return {
+      getI18nText: this.asProps.getI18nText,
+      step: this.asProps.step,
+    };
+  }
+
+  stepperRefs = [];
+  stepperFocusPrev = (i) => () => {
+    const prevStep = this._steps.get(this.asProps.step);
+    if (!prevStep) return;
+    this.setState({ highlighted: prevStep?.step });
+    setTimeout(() => {
+      this.stepperRefs[i - 1]?.focus();
+    }, 0);
+  };
+  stepperFocusNext = (i) => () => {
+    const nextStep = this._steps.get(this.asProps.step + 2);
+    if (!nextStep) return;
+    this.setState({ highlighted: nextStep?.step });
+    setTimeout(() => {
+      this.stepperRefs[i + 1]?.focus();
+    }, 0);
+  };
 
   getStepperProps(props, i) {
     let number = i + 1;
@@ -39,32 +89,45 @@ class WizardRoot extends Component {
     } else {
       this._steps.set(props.step, { number, ...props });
     }
+    const active = props.step === this.asProps.step;
+    const highlighted =
+      this.state.highlighted === props.step || (this.state.highlighted === null && i === 0);
     return {
-      active: props.step === this.asProps.step,
+      active,
+      tabIndex: highlighted ? 0 : -1,
       number,
       getI18nText: this.asProps.getI18nText,
+      uid: this.asProps.uid,
+      ref: (node) => {
+        this.stepperRefs[i] = node;
+      },
+      focusNext: this.stepperFocusNext(i),
+      focusPrev: this.stepperFocusPrev(i),
     };
   }
 
   componentDidUpdate(prevProps) {
     if (prevProps.step === this.asProps.step) return;
+    this.setState({ highlighted: this.asProps.step || null });
     setTimeout(() => {
       if (prevProps.step === this.asProps.step) return;
-      if (document.activeElement !== document.body) return;
-      if (!this.asProps.visible) return;
-      if (!this.modalRef.current) return;
-      setFocus(this.modalRef.current, document.body);
+      setFocus(this.contentRef.current);
     }, 1);
   }
 
   render() {
     const SWizard = this.Root;
-    const { Children, styles } = this.asProps;
+    const { Children, styles, uid } = this.asProps;
 
     this._steps.clear();
 
     return sstyled(styles)(
-      <SWizard render={Modal} ref={this.modalRef}>
+      <SWizard
+        render={Modal}
+        aria-label={undefined}
+        ref={this.modalRef}
+        aria-labelledby={`${uid}-title`}
+      >
         <Children />
       </SWizard>,
     );
@@ -72,13 +135,19 @@ class WizardRoot extends Component {
 }
 
 function Sidebar(props) {
-  const { Children, styles, title } = props;
+  const { Children, styles, title, uid } = props;
   const SSidebar = Root;
   const SSidebarHeader = 'h2';
+  const SSidebarMenu = 'div';
+
+  const handleKeyDown = React.useCallback((e) => {}, []);
+
   return sstyled(styles)(
-    <SSidebar render={Box} role='menu'>
-      {title && <SSidebarHeader>{title}</SSidebarHeader>}
-      <Children />
+    <SSidebar render={Box} __excludeProps={['title']}>
+      {title && <SSidebarHeader id={`${uid}-title`}>{title}</SSidebarHeader>}
+      <SSidebarMenu role='tablist' aria-orientation='vertical'>
+        <Children />
+      </SSidebarMenu>
     </SSidebar>,
   );
 }
@@ -88,7 +157,7 @@ function Step(props) {
   const { Children, styles, active } = props;
   if (active) {
     return sstyled(styles)(
-      <SStep render={Box} tag={Box}>
+      <SStep render={Box}>
         <Children />
       </SStep>,
     );
@@ -97,8 +166,20 @@ function Step(props) {
 }
 
 function Stepper(props) {
-  const { Children, styles, step, active, onActive, completed, disabled, number, getI18nText } =
-    props;
+  const {
+    Children,
+    styles,
+    step,
+    active,
+    onActive,
+    completed,
+    disabled,
+    number,
+    getI18nText,
+    uid,
+    focusNext,
+    focusPrev,
+  } = props;
   const SStepper = Root;
   const SStepNumber = 'span';
   const SStepDescription = 'span';
@@ -117,16 +198,26 @@ function Stepper(props) {
         e.preventDefault();
         onActive(step, e);
       }
+      if (e.key === 'ArrowUp') {
+        focusPrev();
+        e.stopPropagation();
+      }
+      if (e.key === 'ArrowDown') {
+        focusNext();
+        e.stopPropagation();
+      }
     },
-    [step, onActive],
+    [step, onActive, focusPrev, focusNext],
   );
 
   return sstyled(styles)(
     <SStepper
       render={Box}
-      role='menuitem'
+      role='tab'
+      id={`${uid}-stepper-${step}`}
+      aria-controls={active ? `${uid}-content-${step}` : undefined}
       aria-disabled={disabled}
-      aria-current={active}
+      aria-selected={active}
       onClick={handlerClick}
       onKeyDown={handlerKeyDown}
     >
@@ -142,12 +233,60 @@ function Stepper(props) {
 Stepper.enhance = [keyboardFocusEnhance()];
 
 function Content(props) {
-  const { Children, styles } = props;
+  const { Children, children: hasChildren, styles, uid, step } = props;
   const SContent = Root;
   return sstyled(styles)(
-    <SContent render={Box}>
-      <Children />
+    <SContent
+      render={Box}
+      role='tabpanel'
+      aria-labelledby={`${uid}-stepper-${step}`}
+      id={`${uid}-content-${step}`}
+    >
+      {hasChildren ? <Children /> : stepName}
     </SContent>,
+  );
+}
+
+function StepBack(props) {
+  const SStepBack = Root;
+  const { Children, children: hasChildren, styles, getI18nText, stepName } = props;
+  const handleClick = React.useCallback(() => {
+    props.onActive?.(props.step - 1);
+  }, [props.step]);
+  return sstyled(styles)(
+    <SStepBack
+      render={Button}
+      use='tertiary'
+      size='l'
+      onClick={handleClick}
+      aria-label={getI18nText('backButton', { buttonName: stepName })}
+    >
+      <Button.Addon>
+        <ArrowLeft />
+      </Button.Addon>
+      <Button.Text>{hasChildren ? <Children /> : stepName}</Button.Text>
+    </SStepBack>,
+  );
+}
+function StepNext(props) {
+  const SStepNext = Root;
+  const { Children, children: hasChildren, styles, getI18nText, stepName } = props;
+  const handleClick = React.useCallback(() => {
+    props.onActive?.(props.step + 1);
+  }, [props.step]);
+  return sstyled(styles)(
+    <SStepNext
+      render={Button}
+      use='tertiary'
+      size='l'
+      onClick={handleClick}
+      aria-label={getI18nText('nextButton', { buttonName: stepName })}
+    >
+      <Button.Text>{hasChildren ? <Children /> : stepName}</Button.Text>
+      <Button.Addon>
+        <ArrowRight />
+      </Button.Addon>
+    </SStepNext>,
   );
 }
 
@@ -156,6 +295,8 @@ const Wizard = createComponent(WizardRoot, {
   Content,
   Step,
   Stepper,
+  StepBack,
+  StepNext,
 });
 
 export const wrapWizardStepper = (wrapper) => wrapper;