+++ semcore/data-table/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [4.44.0] - 2024-09-09
+## [4.44.0] - 2024-09-20
 
 ### Added